MIDI/Monophonic

+++ Source/EnvelopeDisplay.h
@@ -21,6 +21,7 @@
 ******************************************************************************************************************************/
 
 #pragma once
+#include <cstdint>
 #include <iostream>
 #include <JuceHeader.h>
 #include "PluginProcessor.h"
@@ -36,7 +37,7 @@ class EnvelopeDisplay: public juce::Component
         int   diameter = gap * 2 + 1;
         float opacity = 0.2f;
         float curve_width = 2.0f;
-        const int id;
+        const uint8_t id;
         core::Envelope              env;            // Displayed envelope
         juce::Rectangle<int>        area;           // Display bounds
         core::Rectangle<int>        scope_bounds;   // Scope constraints
